+++ src/algorithms/stats/histogram.cpp
@@ -60,7 +60,7 @@ void Histogram::compute() {
   histogram.resize(_numberBins);
   binEdges.assign(tempBinEdges.begin(), tempBinEdges.end());
 
-  for(int i = 0; i < array.size(); i++){
+  for(size_t i = 0; i < array.size(); i++){
     if(array[i] < _maxValue && array[i] >= _minValue)
       histogram[floor(array[i]/(Real)binWidth)]++;
     else if(array[i] == _maxValue)
